fix(npc): AskMemberShopAvatar candidate-count is a byte, not int
CScriptMan::OnAskMembershopAvatar@0x6dd340 (case 9 of
CScriptMan::OnScriptMessage) reads the candidate count via
CInPacket::Decode1 (1 byte) at 0x6dd394 before looping Decode4 per
candidate. Atlas wrote the count as WriteInt (4 bytes), desyncing the
client's per-candidate id reads. Switch to WriteByte.

+++ libs/atlas-packet/npc/clientbound/conversation.go
@@ -268,7 +268,7 @@ func (a *AskMemberShopAvatarConversationDetail) Encode(l logrus.FieldLogger, _ c
 	w := response.NewWriter(l)
 	return func(options map[string]interface{}) []byte {
 		w.WriteAsciiString(a.Message)
-		w.WriteInt(uint32(len(a.Candidates)))
+		w.WriteByte(byte(len(a.Candidates)))
 		for _, candidate := range a.Candidates {
 			w.WriteInt(candidate)
 		}
